remove BouncyCastle dependency
Since the BouncyCastle bcprov artefact is no longer a Keycloak dependency, the plugin has to ensure that it has an MD4 implementation itself.

In order not to introduce BouncyCastle as an additional security provider into the entire Keycloak JVM, an MD4 implementation from org.eclipse.angus:angus-mail is used, which is already on the classpath as a dependency of keycloak-services.